python-backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "AagamChhajer",email = "aagamchhajer@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.9"
dependencies = [
    "boto3 (>=1.37.4,<2.0.0)",
    "botocore (>=1.37.4,<2.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.10,<0.116.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-magic-bin (>=0.4.14,<0.5.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "whisper-timestamped (>=1.15.8,<2.0.0)",
    "moviepy (>=2.1.2,<3.0.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "google-cloud-texttospeech (>=2.25.0,<3.0.0)",
    "google-cloud-aiplatform (>=1.82.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
